+++ src/main/resources/mapper/attr.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
 	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
  
-<mapper namespace="com.ssafy.enjoytrip.attr.model.dao.AttrInfoDao">
+<mapper namespace="com.ssafy.enjoytrip.attr.model.mapper.AttrInfoMapper">
 
 
 	<resultMap type="SidoDto" id="sidoList">
